prototype/pyproject.toml
[project]
name = "hackathon-agent"
version = "0.1.0"
description = "Minimal schemas for the Agents Assemble hackathon demo"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1",
  "google-genai>=1.30,<2",
  "pydantic>=2.7,<3",
  "uvicorn>=0.30,<1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
